+++ SpringBoot01/src/main/java/com/example/studyspringboot/domain/feed/presentation/FeedController.java
@@ -29,7 +29,7 @@ public class FeedController {
     }
 
     @GetMapping("/{feed-id}")
-    public FeedDetailResponse getFeedDetail(@PathVariable("feed-id") Long feedId) {
+    public FeedDetailResponse getFeedDetail(@PathVariable("feed-id") Integer feedId) {
         return queryFeedDetailService.getFeedDetail(feedId);
     }
